Extract uri convertor
