+++ src/ProductConstructionService/ProductConstructionService.Api/Api/v2020_02_20/Controllers/PullRequestController.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Net;
-using System.Text.RegularExpressions;
 using Maestro.Common;
 using Maestro.Data;
 using Microsoft.AspNetCore.ApiVersioning;
@@ -21,29 +20,6 @@ namespace ProductConstructionService.Api.Api.v2020_02_20.Controllers;
 [ApiVersion("2020-02-20")]
 public partial class PullRequestController : ControllerBase
 {
-    [GeneratedRegex(@"https://api.github.com/repos/(?<org>[^/]+)/(?<repo>[^/]+)/pulls/(?<id>[0-9]+)/?")]
-    private static partial Regex GitHubApiPrUrlRegex();
-
-    [GeneratedRegex(@"https://dev.azure.com/(?<org>[^/]+)/(?<project>[^/]+)/_apis/git/repositories/(?<repo>[^/]+)/pullRequests/(?<id>[0-9]+)/?")]
-    private static partial Regex AzdoApiPrUrlRegex();
-
-    private static readonly Dictionary<string, string> WellKnownIds = new()
-    {
-        ["7ea9116e-9fac-403d-b258-b31fcf1bb293"] = "internal", // https://dev.azure.com/dnceng/internal
-        ["0bdbc590-a062-4c3f-b0f6-9383f67865ee"] = "DevDiv", // https://dev.azure.com/devdiv/DevDiv
-        ["55e8140e-57ac-4e5f-8f9c-c7c15b51929d"] = "ProjectReunion", // https://dev.azure.com/microsoft/ProjectReunion
-    };
-
-    private static string ResolveWellKnownIds(string str)
-    {
-        foreach (var pair in WellKnownIds)
-        {
-            str = str.Replace(pair.Key, pair.Value);
-        }
-
-        return str;
-    }
-
     private readonly IRedisCacheFactory _cacheFactory;
     private readonly BuildAssetRegistryContext _context;
 
@@ -104,7 +80,7 @@ public partial class PullRequestController : ControllerBase
 
             var updates = subscriptions
                 .Select(update => new PullRequestUpdate(
-                    TurnApiUrlToWebsite(update.SourceRepository, org, repoName),
+                    GitRepoUrlUtils.TurnApiUrlToWebsite(update.SourceRepository, org, repoName),
                     pr.ContainedSubscriptions.First(s => s.SubscriptionId == update.Id).SubscriptionId,
                     pr.ContainedSubscriptions.First(s => s.SubscriptionId == update.Id).BuildId))
                 .ToList();
@@ -178,32 +154,6 @@ public partial class PullRequestController : ControllerBase
         return await cache.TryDeleteAsync() == null ? NotFound() : Ok();
     }
 
-    private static string TurnApiUrlToWebsite(string url, string? orgName, string? repoName)
-    {
-        var match = GitHubApiPrUrlRegex().Match(url);
-        if (match.Success)
-        {
-            return $"https://github.com/{match.Groups["org"]}/{match.Groups["repo"]}/pull/{match.Groups["id"]}";
-        }
-
-        match = AzdoApiPrUrlRegex().Match(url);
-        if (match.Success)
-        {
-            // If we have the repo name, use it to replace the repo GUID in the URL
-            if (repoName != null)
-            {
-                WellKnownIds[match.Groups["repo"].Value] = orgName + "-" +repoName;
-            }
-
-            var org = ResolveWellKnownIds(match.Groups["org"].Value);
-            var project = ResolveWellKnownIds(match.Groups["project"].Value);
-            var repo = ResolveWellKnownIds(match.Groups["repo"].Value);
-            return $"https://dev.azure.com/{org}/{project}/_git/{repo}/pullrequest/{match.Groups["id"]}";
-        }
-
-        return url;
-    }
-
     internal static TrackedPullRequest ToTrackedPullRequest(
         InProgressPullRequest pr,
         string subscriptionId,
@@ -212,7 +162,7 @@ public partial class PullRequestController : ControllerBase
         string? org = null;
         string? repoName = null;
         // Only extract org/repoName for Azure DevOps PRs
-        if (subscription != null && AzdoApiPrUrlRegex().IsMatch(pr.Url))
+        if (subscription != null && GitRepoUrlUtils.IsAzdoApiPrUrl(pr.Url))
         {
             try
             {
@@ -226,7 +176,7 @@ public partial class PullRequestController : ControllerBase
 
         return new TrackedPullRequest(
             subscriptionId,
-            TurnApiUrlToWebsite(pr.Url, org, repoName),
+            GitRepoUrlUtils.TurnApiUrlToWebsite(pr.Url, org, repoName),
             subscription?.Channel != null ? new Channel(subscription.Channel) : null,
             subscription?.TargetBranch,
             subscription?.SourceEnabled ?? false,
