+++ test/Darc/Microsoft.DotNet.DarcLib.Tests/Helpers/GitRepoUrlParserTest.cs
@@ -97,4 +97,33 @@ public class GitRepoUrlParserTest
     {
         GitRepoUrlUtils.GetCommitUri(repoUri, commit).Should().Be(expectedUri);
     }
+
+    [Test]
+    [TestCase("https://api.github.com/repos/dotnet/dotnet/pulls/3205", "https://github.com/dotnet/dotnet/pull/3205")]
+    [TestCase("https://api.github.com/repos/dotnet/runtime/pulls/12345", "https://github.com/dotnet/runtime/pull/12345")]
+    [TestCase("https://api.github.com/repos/microsoft/CsWinRT/pulls/999", "https://github.com/microsoft/CsWinRT/pull/999")]
+    public void TurnApiUrlToWebsite_ConvertsGitHubApiUrlsToWebUrls(string apiUrl, string expectedWebUrl)
+    {
+        var result = GitRepoUrlUtils.TurnApiUrlToWebsite(apiUrl, null, null);
+        result.Should().Be(expectedWebUrl);
+    }
+
+    [Test]
+    [TestCase("https://dev.azure.com/dnceng/7ea9116e-9fac-403d-b258-b31fcf1bb293/_apis/git/repositories/test-repo-guid/pullRequests/123",
+              "https://dev.azure.com/dnceng/internal/_git/test-repo-guid/pullrequest/123")]
+    public void TurnApiUrlToWebsite_ConvertsAzureDevOpsApiUrlsToWebUrls(string apiUrl, string expectedWebUrl)
+    {
+        var result = GitRepoUrlUtils.TurnApiUrlToWebsite(apiUrl, null, null);
+        result.Should().Be(expectedWebUrl);
+    }
+
+    [Test]
+    [TestCase("https://github.com/dotnet/dotnet/pull/3205")]
+    [TestCase("https://dev.azure.com/dnceng/internal/_git/dotnet-wpf/pullrequest/123")]
+    [TestCase("not-a-url")]
+    public void TurnApiUrlToWebsite_ReturnsOriginalUrlWhenNotApiUrl(string url)
+    {
+        var result = GitRepoUrlUtils.TurnApiUrlToWebsite(url, null, null);
+        result.Should().Be(url);
+    }
 }
